создал FilmsController
